Fetch site base from configuration instead of getter
The TYPO3 site handling has a "base" property which is the domain
of the website. But there are baseVariants that are applied if a
condition is matched and also languages that may overwrite the base.
That means the "base" is not really consistend and does not really
identify the "site".
Instead of using the generated "base" by using the getter the raw
"base" configuration value is used to identify the site.

Resolves: #2578